Introduce a JoystickCameraController with adjustable sensitivity.
The "sensitivity" is a parameter that determines the VISCA output value
when the joystick is at 50% position. Output values are determined by
linear interpolation based on that point and the min and max values.

Setting sensitivity to 50% of the maximum VISCA output value, then, will
give a linear response across the entire joystick axis range, and is the
default (based on Sony VISCA speed limits for now).