Debounce buttons.
It will no longer be possible to accidentally double-tap a button thanks
to the vagiaries of a touchscreen input device; each button can be
clicked at most once every quarter of a second. We might want to tweak
that parameter if it becomes annoying (or noticable).

Relates to, but does not fix, #34.